kernel: move chip specific print to arch crate
